site/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==4.3.0
django-widget-tweaks==1.5.0
djangorestframework==3.13.1
python-dotenv==0.21.0
pytz==2023.3.post1
sqlparse==0.4.3
